requirements.txt
python-dotenv==1.0.0
langchain==0.0.353
openai==1.6.1
chromadb==0.4.21
tiktoken==0.5.2
duckduckgo-search==4.1.1
numexpr==2.8.8
wikipedia==1.4.0
wolframalpha==5.0.0
